requirements.txt
langchain
langchain-community
chromadb
pypdf
python-docx
beautifulsoup4
requests
sentence-transformers
ollama